requirements.txt
flask==2.1.2
pymongo==4.1.1
mysql-connector==2.2.9